[Spiders][Not Working] Connect the genius spider and django using py script
This commit connects the spiders and the django server using an external
python script.

The method used is based off of scrapy documentation and
twisted reactors, this is not an ideal way to do it but since things have
been configured this commit will server as a baseline for the next approach
to establish the connection using scrapyd. Hosting spiders on another port.